requirements.txt
python_decouple==3.1
dj_database_url==0.4.2
Django==2.0.1
djangorestframework==3.7.7
requests==2.18.4
django-widget-tweaks==1.4.1
gunicorn==19.7.1
psycopg2==2.7.3.2
whitenoise==3.3.1
